refactored api users to separate from users_auth endpoints for organizational purposes
