+++ api/routes/index.js
@@ -1,9 +1,11 @@
-var express = require("express");
-var router = express.Router();
+// not needed for development/production
 
-/* GET home page. */
-router.get("/auth0", function(req, res, next) {
-  res.render("index", { title: "Auth0 for Auto Invoicer App" });
-});
+// var express = require("express");
+// var router = express.Router();
 
-module.exports = router;
+// /* GET home page. */
+// router.get("/auth0", function(req, res, next) {
+//   res.render("index", { title: "Auth0 for Auto Invoicer App" });
+// });
+
+// module.exports = router;
