--- api/routes/login.js
@@ -1,7 +1,0 @@
-const router = require("express").Router();
-
-router.post("/", (req, res) => {
-  // for authenticating existing users
-});
-
-module.exports = router;
